+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
 ### Improvements
 
+- Avoid assigning the Shift+Return shortcut key for a snippet.
 - [trivial] Update find result messages.
 - [dev] Update the build environment to Xcode 26.0.
 - [non-AppStore ver.] Update Sparkle from 2.8.0-beta.3 to 2.8.0.